requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
